requirements.txt
airtest==1.1.4
certifi==2020.6.20
chardet==3.0.4
comtypes==1.1.7
hrpc==1.0.8
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
mss==4.0.3
numpy==1.19.1
opencv-contrib-python==3.4.2.17
Pillow==7.2.0
pocoui==1.0.79
pypiwin32==223
pywin32==228
pywinauto==0.6.3
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.48.0
